indexer-service/requirements.txt
Flask
gunicorn
google-cloud-discoveryengine
